feat: super-linter support
This adds a GitHub workflow that runs super-linter on all PRs. It
configures super-linter using the super-linter.env file which sets env
vars for the linter. It also adds a script to run the linter locally
using the same env vars. A VSCode task was added which runs the script.

Finally, a .editorconfig file was added to ensure the linting and editor
settings are the same.

Refs: #22
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>